Specify nonces as 64-bit in Go implementations

+++ src/go/1types.go
@@ -15,7 +15,7 @@ type messagebuffer struct {
 
 type cipherstate struct {
 	k [32]byte
-	n uint32
+	n uint64
 }
 
 type symmetricstate struct {
